+++ examples/echo.rs
@@ -5,7 +5,7 @@ fn main() {
     let caprice = Caprice::new()
         .set_prompt("!:") // set the prompt
         .disable_ctrl_c() // pressing control + c won't terminate the caprice console
-        .set_keywords(&[
+        .set_keywords(vec![
             // set some tokens
             "some_token".to_owned(),
             "some_other_token".to_owned(),